+++ src/main/kotlin/de/jfachwert/FachwertFactory.kt
@@ -36,6 +36,7 @@ import de.jfachwert.util.SmallUUID
 import de.jfachwert.util.TinyUUID
 import de.jfachwert.zeit.Zeitdauer
 import de.jfachwert.zeit.Zeitpunkt
+import de.jfachwert.zeit.Zeitraum
 import java.io.Serializable
 import java.lang.reflect.InvocationTargetException
 import java.util.*
@@ -196,6 +197,7 @@ class FachwertFactory private constructor() {
             instance.register(PZN::class.java)
             instance.register(Zeitdauer::class.java)
             instance.register(Zeitpunkt::class.java)
+            instance.register(Zeitraum::class.java)
         }
     }